refactor: change services to service
