+++ service/jwt.go
@@ -1,4 +1,4 @@
-package services
+package service
 
 import (
 	"fmt"
